requirements.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
Appium-Python-Client==1.2.0
argcomplete==1.12.3
atomicwrites==1.4.0
attrs==20.3.0
backports.entry-points-selectable==1.1.0
blinker==1.4
Brotli==1.0.9
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==7.1.2
colorama==0.4.4
colorlog==4.8.0
cryptography==2.9.2
distlib==0.3.2
enum34==1.1.10
filelock==3.0.12
filetype==1.0.7
Flask==1.1.4
gitdb==4.0.5
GitPython==3.1.12
h11==0.12.0
h2==3.2.0
har2case==0.3.1
hpack==3.0.0
httprunner==2.5.7
hyperframe==5.2.0
idna==3.2
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonpath==0.82
jsonschema==3.2.0
kaitaistruct==0.8
ldap3==2.7
lxml==4.6.2
MarkupSafe==2.0.1
mitmproxy==5.2
namedlist==1.8
packaging==20.9
passlib==1.7.4
pipx==0.16.3
platformdirs==2.0.2
pluggy==0.13.1
protobuf==3.11.3
publicsuffix2==2.20191221
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
pydivert==2.1.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.2
pyrsistent==0.18.0
pytest==6.2.2
pytest-rerun==0.0.1
pytest-rerunfailures==9.1.1
pytest-sugar==0.9.4
pytest-testconfig==0.2.0
PyYAML==5.4.1
requests==2.26.0
requests-toolbelt==0.9.1
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.6
selenium==3.141.0
sentry-sdk==0.13.5
six==1.15.0
smmap==3.0.5
sortedcontainers==2.1.0
termcolor==1.1.0
toml==0.10.2
tornado==6.1
urllib3==1.26.6
urwid==2.0.1
userpath==1.6.0
virtualenv==20.6.0
Werkzeug==1.0.1
wsproto==0.15.0
zstandard==0.13.0
